MAI_v_3.0: выполнен рефакторинг кода: приведение к Style Guide: - добавлены default в case - замена в SetText("") "" на EMPTY_NAME - обработаны exceptions - убрана глобальная переменная - исправлена область видимости полей - имена полей приведены в соответствии с Android Code Style

+++ MAI/Search-Person-Popularity-MAI/app/src/main/java/ru/geekbrains/traineeship/group2/search_person_popularity_mai/Repository/Data/Person.java
@@ -70,7 +70,6 @@ public class Person {
 
     @Override
     public int hashCode() {
-        return (int) id *
-                name.hashCode();
+        return id * name.hashCode();
     }
 }
