+++ MAI/Search-Person-Popularity-MAI/app/src/main/java/ru/geekbrains/traineeship/group2/search_person_popularity_mai/Repository/Players/Admin.java
@@ -115,7 +115,7 @@ public class Admin {
 
     @Override
     public int hashCode() {
-        return (int) id *
+        return id *
                 nickName.hashCode() *
                 login.hashCode();
     }
